Fix 'new' compile error. Pass a parameter either like primitive type ( byte, short...) or null it's an error. It's safe pass parameter either like Short, Byte ... or null

+++ aon-db-up2date/src/main/java/net/aonsolutions/db/up2date/fiscal/RefreshMod123Result.java
@@ -93,7 +93,7 @@ public class RefreshMod123Result implements Update {
 							if (result != null) {
 								dslContext.update(FS_MODEL)
 									.set(FS_MODEL.RESULT,result)
-									.set(FS_MODEL.DECLARATION_TYPE, (declarationType == null? null : declarationType.value()))
+									.set(FS_MODEL.DECLARATION_TYPE, (declarationType == null? null : (Byte)declarationType.value()))
 									.where(FS_MODEL.ID.eq(id))
 									.execute();
 							}
